Add SetSearchTextInterceptor and improve data fetching
A new class SetSearchTextInterceptor is added to automatically create a search text while saving changes to the database. This interceptor is added to the DbContext's options. Moreover, data fetching in the CatalogService is optimized by adding .AsNoTracking() to prevent unnecessary change tracking and by elaborating .Include() to fetch all necessary related data.