Add search functionality to the catalog
A new search field has been added to the catalog which allows users to search for items dynamically. This functionality uses a timer to delay the search until the user has finished typing. Additionally, an API call has been added to the catalog service to fetch search results. A clear search function was also introduced to reset the search field.